requirements.txt
bin2coe
PyQt5
Pillow
